requirements.txt
deepl==1.18.0
PyYAML==6.0.2
